Release the PKCS11_ structures when looking for RSA keys
Basically we register an index on the RSA structure to be able to associate
the PKCS11_ data with the RSA key.
The destructor of the registered index takes care of deleting the PKCS11_
data using p11 PKCS11_release_all_slots.
The key used is a special case, but NULLing the evp_key of it, prevents p11
from calling EVP_PKEY_free on it, else we'd hit recursion, having
EVP_PKEY_free calling RSA_free calling the destructor again.

A problem when using the index is - it can not be unregistered.
So un-loading the engine will leave an invalid callback.
Therefore - better not un-load the engine.