Default to using p11-kit-proxy.so if it's available
In the general case, PKCS#11-aware applications should automatically be
using the tokens listed in the system's p11-kit configuration, rather than
needing to be explicitly pointed at a provider module.

We currently just *fail* if no provider is given. Fix that to use the
p11-kit-proxy.so module from p11-kit if it's detected.

Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>